backend/service/requirements.txt
cmake
asgiref
bidict
certifi
click
Flask
Flask-Cors
Flask-SocketIO
image
imutils
itsdangerous
Jinja2
MarkupSafe
numpy
Pillow
pipenv
python-engineio
python-socketio
pytz
six
sqlparse
virtualenv
virtualenv-clone
Werkzeug
opencv-python
dlib
scipy
flask-sqlalchemy
flask-migrate
face_recognition
mtcnn
tensorflow